Connecting Like and Comment with login orch and art collection service

+++ like_comment_management/like_comment/views.py
@@ -31,9 +31,12 @@ def get_user(request):
         try:
             data = response.json()
             username = data['username']
-            customer = Cust(username=username)
-            customer.save()
-            return username
+            if not (Cust.objects.filter(username=username).exists()) :
+                customer = Cust(username=username)
+                customer.save()
+                return username
+            else :
+                return username
         except json.decoder.JSONDecodeError as e:
             # Handle the JSONDecodeError
             return JsonResponse({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
@@ -47,48 +50,58 @@ def get_collection_id(request):
     return "http://localhost:8084"
 
 @csrf_exempt
-def get_post(request, post_id):
-    post = Collection.objects.filter(post_id=post_id).values()[0]
-    return JsonResponse({"response":post}, status=status.HTTP_200_OK)
+def get_post(request, collection_id):
+    collection = Collection.objects.filter(collection_id=collection_id).values()[0]
+    return JsonResponse({"response":collection}, status=status.HTTP_200_OK)
 
 @csrf_exempt
-def like(request, post_id):
+def like(request, collection_id):
     user = get_user(request)
-    post = Collection.objects.get_or_create(post_id=post_id)
-    like = Like.objects.filter(like_post=post, like_user=user)
-    if not like.exists():
-        add_like = Like(like_post=post, like_user=user)
-        add_like.save()
-        post.post_likes += 1
-        post.save()
-        return JsonResponse({'message': 'Post Liked'}, status=status.HTTP_200_OK)
-    else:
-        like.delete()
-        post.post_likes -= 1
-        post.save()
-        return JsonResponse({'message': 'Post Unliked'}, status=status.HTTP_200_OK)
-    
+    filter_collection = Collection.objects.filter(collection_id = collection_id).exists() 
+    if filter_collection == True :
+        collection = Collection.objects.get(collection_id=collection_id)
+        like = Like.objects.filter(like_collection=collection, like_user=user)
+        if not like.exists():
+            add_like = Like(like_collection=collection, like_user=user)
+            add_like.save()
+            collection.collection_likes += 1
+            collection.save()
+            return JsonResponse({'message': 'Post Liked'}, status=status.HTTP_200_OK)
+        else:
+            like.delete()
+            collection.collection_likes -= 1
+            collection.save()
+            return JsonResponse({'message': 'Post Unliked'}, status=status.HTTP_200_OK)
+    else :
+        create_collection = Collection(collection_id = collection_id)
+        create_collection.save()
+        return JsonResponse({'message' : 'Collection Created Successfully'}, status=status.HTTP_200_OK)
+        
 @csrf_exempt
-def comment(request, post_id):
+def comment(request, collection_id):
     user = get_user(request)
-    post = Collection.objects.get(post_id = post_id)
-    deserialize = json.loads(request.body)
-    comment = Comment(comment_post=post, comment_user=user, comment_text=deserialize['comment'])
-    comment.save()
-    return JsonResponse({'message': 'Comment added successfully'}, status=status.HTTP_200_OK)
+    filter_collection = Collection.objects.filter(collection_id = collection_id).exists() 
+    if filter_collection == True :
+        collection = Collection.objects.get(collection_id = collection_id)
+        deserialize = json.loads(request.body)
+        comment = Comment(comment_collection=collection, comment_user=user, comment_text=deserialize['comment'])
+        comment.save()
+        return JsonResponse({'message': 'Comment added successfully'}, status=status.HTTP_200_OK)
+    else :
+        create_collection = Collection(collection_id = collection_id)
+        create_collection.save()
+        return JsonResponse({'message' : 'Collection Created Successfully'}, status=status.HTTP_200_OK)
 
 @csrf_exempt
-def get_comment(request, post_id):
-    post = Collection.objects.get(post_id = post_id)
-    comment_list = Comment.objects.filter(comment_post=post).order_by('-comment_id').values()[::1]
+def get_comment(request, collection_id):
+    collection = Collection.objects.get(collection_id = collection_id)
+    comment_list = Comment.objects.filter(comment_collection=collection).order_by('-comment_id').values()[::1]
 
     return JsonResponse({"response":comment_list}, status=status.HTTP_200_OK)
 
-def liked(request, post_id):
-    result = Collection.object.filter(post_id=post_id, post_likes_user=request.user.email)
-    if len(result) == 0:
-        is_liked = False
-    else :
-        is_liked = True
-    return JsonResponse({"isLiked":is_liked}, safe=False, status=status.HTTP_200_OK)
+@csrf_exempt
+def liked_by_user(request):
+    user = get_user(request)
+    result = Like.objects.filter(like_user = user).order_by('-like_collection').values()[::1]
+    return JsonResponse({"response":result}, status=status.HTTP_200_OK)
